+++ examples/grpc/tracer_common.h
@@ -1,0 +1,91 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+#include "opentelemetry/exporters/ostream/span_exporter.h"
+#include "opentelemetry/sdk/trace/simple_processor.h"
+#include "opentelemetry/sdk/trace/tracer_provider.h"
+#include "opentelemetry/trace/provider.h"
+
+#include "opentelemetry/context/propagation/global_propagator.h"
+#include "opentelemetry/context/propagation/text_map_propagator.h"
+#include "opentelemetry/nostd/shared_ptr.h"
+#include "opentelemetry/trace/propagation/http_trace_context.h"
+
+#include <grpcpp/grpcpp.h>
+#include <cstring>
+#include <iostream>
+#include <vector>
+
+using grpc::ClientContext;
+using grpc::ServerContext;
+
+namespace
+{
+class GrpcClientCarrier : public opentelemetry::context::propagation::TextMapCarrier
+{
+public:
+  GrpcClientCarrier(ClientContext *context) : context_(context) {}
+  GrpcClientCarrier() = default;
+  virtual nostd::string_view Get(nostd::string_view key) const noexcept override { return ""; }
+
+  virtual void Set(nostd::string_view key, nostd::string_view value) noexcept override
+  {
+    std::cout << " Client ::: Adding " << key << " " << value << "\n";
+    context_->AddMetadata(key.data(), value.data());
+  }
+
+  ClientContext *context_;
+};
+
+class GrpcServerCarrier : public opentelemetry::context::propagation::TextMapCarrier
+{
+public:
+  GrpcServerCarrier(ServerContext *context) : context_(context) {}
+  GrpcServerCarrier() = default;
+  virtual nostd::string_view Get(nostd::string_view key) const noexcept override
+  {
+    auto it = context_->client_metadata().find(key.data());
+    if (it != context_->client_metadata().end())
+    {
+      return it->second.data();
+    }
+    return "";
+  }
+
+  virtual void Set(nostd::string_view key, nostd::string_view value) noexcept override
+  {
+    // Not required for server
+  }
+
+  ServerContext *context_;
+};
+
+void initTracer()
+{
+  auto exporter = std::unique_ptr<sdktrace::SpanExporter>(
+      new opentelemetry::exporter::trace::OStreamSpanExporter);
+  auto processor = std::unique_ptr<sdktrace::SpanProcessor>(
+      new sdktrace::SimpleSpanProcessor(std::move(exporter)));
+  std::vector<std::unique_ptr<sdktrace::SpanProcessor>> processors;
+  processors.push_back(std::move(processor));
+  // Default is an always-on sampler.
+  auto context  = std::make_shared<sdktrace::TracerContext>(std::move(processors));
+  auto provider = nostd::shared_ptr<opentelemetry::trace::TracerProvider>(
+      new sdktrace::TracerProvider(context));
+  // Set the global trace provider
+  opentelemetry::trace::Provider::SetTracerProvider(provider);
+
+  // set global propagator
+  opentelemetry::context::propagation::GlobalTextMapPropagator::SetGlobalPropagator(
+      nostd::shared_ptr<opentelemetry::context::propagation::TextMapPropagator>(
+          new opentelemetry::trace::propagation::HttpTraceContext()));
+}
+
+nostd::shared_ptr<opentelemetry::trace::Tracer> get_tracer(std::string tracer_name)
+{
+  auto provider = opentelemetry::trace::Provider::GetTracerProvider();
+  return provider->GetTracer(tracer_name);
+}
+
+}  // namespace