+++ exporters/elasticsearch/include/opentelemetry/exporters/elasticsearch/es_log_recordable.h
@@ -2,13 +2,14 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #pragma once
+#ifdef ENABLE_LOGS_PREVIEW
 
-#include <map>
-#include <unordered_map>
-#include "nlohmann/json.hpp"
-#include "opentelemetry/sdk/common/attribute_utils.h"
-#include "opentelemetry/sdk/logs/recordable.h"
-#include "opentelemetry/version.h"
+#  include <map>
+#  include <unordered_map>
+#  include "nlohmann/json.hpp"
+#  include "opentelemetry/sdk/common/attribute_utils.h"
+#  include "opentelemetry/sdk/logs/recordable.h"
+#  include "opentelemetry/version.h"
 
 OPENTELEMETRY_BEGIN_NAMESPACE
 namespace exporter
@@ -53,6 +54,9 @@ private:
         return;
       case common::AttributeType::kTypeDouble:
         json_[name][key.data()] = opentelemetry::nostd::get<double>(value);
+        return;
+      case common::AttributeType::kTypeCString:
+        json_[name][key.data()] = opentelemetry::nostd::get<const char *>(value);
         return;
       case common::AttributeType::kTypeString:
         json_[name][key.data()] =
@@ -158,3 +162,4 @@ public:
 }  // namespace logs
 }  // namespace exporter
 OPENTELEMETRY_END_NAMESPACE
+#endif