+++ exporters/etw/test/etw_tracer_test.cc
@@ -12,9 +12,7 @@
 
 using namespace OPENTELEMETRY_NAMESPACE;
 
-using Properties       = opentelemetry::exporter::etw::Properties;
-using PropertyValue    = opentelemetry::exporter::etw::PropertyValue;
-using PropertyValueMap = opentelemetry::exporter::etw::PropertyValueMap;
+using namespace opentelemetry::exporter::etw;
 
 std::string getTemporaryValue()
 {
@@ -60,7 +58,6 @@ TEST(ETWTracer, TracerCheck)
   {
     auto topSpan = tracer->StartSpan("MySpanTop");
     auto topScope = tracer->WithActiveSpan(topSpan);
-    std::this_thread::sleep_for (std::chrono::seconds(1));
     {
       auto outerSpan = tracer->StartSpan("MySpanL2", attribs);
       auto outerScope = tracer->WithActiveSpan(outerSpan);
@@ -80,7 +77,6 @@ TEST(ETWTracer, TracerCheck)
           {"strKey", "someValue"}
         };
         EXPECT_NO_THROW(outerSpan->AddEvent(eventName1, event1));
-        std::this_thread::sleep_for (std::chrono::seconds(1));
 
         // Add second event
         std::string eventName2 = "MyEvent2";
@@ -91,7 +87,6 @@ TEST(ETWTracer, TracerCheck)
           {"strKey", "anotherValue"}
         };
         EXPECT_NO_THROW(outerSpan->AddEvent(eventName2, event2));
-        std::this_thread::sleep_for (std::chrono::seconds(2));
 
         std::string eventName3= "MyEvent3";
         Properties event3 =
@@ -104,7 +99,6 @@ TEST(ETWTracer, TracerCheck)
           {"tempString", getTemporaryValue() }
         };
         EXPECT_NO_THROW(innerSpan->AddEvent(eventName3, event3));
-        std::this_thread::sleep_for (std::chrono::seconds(1));
         EXPECT_NO_THROW(innerSpan->End());
 
       }