+++ sdk/src/trace/batch_span_processor.cc
@@ -116,6 +116,7 @@ void BatchSpanProcessor::DoBackgroundWork()
       // mechanism effort here.
       if (buffer_.empty() == true)
       {
+        timeout = schedule_delay_millis_;
         continue;
       }
     }